requirements.txt
asgiref==3.2.10
Django==3.0.10
django-clearcache==1.0.1
django-crispy-forms==1.9.2
django-forms-ajaxified==0.2
django-libs==2.0.3
django-shortuuidfield==0.1.3
djangorestframework==3.11.1
mysqlclient==2.0.1
pytz==2020.1
shortuuid==1.0.1
six==1.15.0
sqlparse==0.3.1
